+++ js/util.js
@@ -1,0 +1,24 @@
+const getRandomPositiveInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomPositiveFloat = (min, max, digits = 1) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  const result = Math.random() * (upper - lower) + lower;
+  return result.toFixed(digits);
+};
+
+// Функция копирует, затем перемешивает массив в случайном порядке
+const getShuffledItems = (items) => items.slice().sort(() => Math.random() - 0.5);
+
+// Функция возвращает перемешенный массив до заданной длины массива
+const getRandomItems = (items, count) => {
+  const shuffledItems = getShuffledItems(items);
+  return shuffledItems.slice(0, count);
+};
+
+export { getRandomPositiveInteger, getRandomPositiveFloat, getRandomItems };
